project notebook
E-commerce platform becomes reality of disruption that changes the form of physical transactions into online (despite social media), especially in the developing world. Using this platform, entrepreneurs can take advantage of it in business optimization, accelerating business development, access to new market segments, etc. But the question likely to appear is about how to do so?